Removed utils library, added 'FIlename' as field, added caching
Caching is based on last modified time of file. If file is recently modified (<5 days), then it is rescanned. Otherwise, it is obtained from most recent report (i.e.: take from cache). Files which do not reside in report, will obviously be scanned.

Now roughly drops to 3 seconds. Previous implementation which is using multi-threading to scan all files takes 10 seconds.

Of course, could also track last cached datetime, then use that as basis to determine if a file could be obtained from caching, as opposed to looking at hard-coded 5 day which still incurs repeated scanings. But that will perhaps be a future improvement.